+++ CHANGELOG.md
@@ -255,3 +255,4 @@
 * Implement delete method for Stack
 * Define `Comprehension` wrapper for data structures that can behave like list comprehensions
 * Define experimental `Insertable` typeclass
+* Define `Day convolution` datatype in `Transformer` module